Ensure DB connections are checked in
Using PB's around service filters, wrap each request in an AR
`with_connection` call so that connections are checked in and out with
each RPC request. This will ensure that DB connections are checked in
after each RPC request and (hopefully) make PB::AR more compatibile with
Rails 4 (which no longer ensures connections are checked in for you).