Allow attributes to be set to nil
This adds logic to check a repeated string field called 'nullify'
on incoming proto messages.  If there is an attribute with the same name
as an attribute specified in the nullify field, this attribute will be
set to nil.

For transformed attributes there is now an option when the transformer
is declared that will specify a name in the nullify field to inspect.
If the our key is in the nullify array of an incoming proto, the
transformer will set the transformed attribute to nil.