Save protobuf mappings at class level
Calculating protobuf mappings is expensive and requires looping over all
fields and checking respond_to.  We can memoize the field to attribute
mappings.

I had to move the application of the serializaiton options last so that
we could memoize the mapped fields and then apply the options later,
instead of applying the options first.