Fix nested attributes support
It turns out that nested attributes support was added, but never quite
worked (specs would have caught that **sigh**). There were a couple of
missing pieces: two methods needed to be extended to convert Protobuf
messages to attributes, and nested attribute fields needed to be
extracted from the Protobuf messages.

Both have been implemented, and specs have been added. A lot of specs
needed to be updated in the process because changing the sample Protobuf
messages altered their expectations.