Fetch records when serializing collections
Until `to_a` is called on a collection association, it remains a
collection proxy so that it can be used with ARel scopes. This is
problematic when attempting to serialize collection associations to
protobuf messages since they do not respond to `to_proto`.

Add a writer method that terminates collections for fields that are
being serialized from collection associations.